Add files via upload
In this projet we will be using python open cv and xlwrite 

here open cv helps capture the faces of students 1 at a time 
it can be done roll no. wise 
and it is compared to the harr cascade and stored in a dataset folder 
the image is stored with the name user 1.1 1 represents roll no.1 and the no. after the . says the amount of images of the student 

from there it is trained and then later recognizes the face of the student 
and the recognized students name is updated in an excel sheet